requirements.txt
python-dotenv
kiteconnect==5.0.1
pydantic
# mcp-sdk